refactor(db): move moderation store ownership
